LTK affilate/requirements.txt
python-dotenv==1.0.0
supabase==2.3.5
python-jose==3.3.0 